Began Script to Counter Lengthy Study Titles
My idea was to reduce the font-size for the .event h3 if a Bible
study’s title is too long, such as “Cross-Examination: The Evidence for
Belief”, but this isn’t the best solution. I’ll work on it some more
later — we might have to just stick “…” in there. If you have any
ideas, feel free to mess with it.